Service migrated to CrudService

+++ src/main/java/com/parameters/prts/Repository/LiteratureRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LiteratureRepository extends JpaRepository<LiteratureEntity, Long> {
+public interface LiteratureRepository extends JpaRepository<LiteratureEntity, Integer> {
 }
