+++ src/main/java/com/parameters/prts/Repository/PrimaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PrimaryRepository extends JpaRepository<PrimaryEntity, Long> {
+public interface PrimaryRepository extends JpaRepository<PrimaryEntity, Integer> {
 }
