+++ src/main/java/com/parameters/prts/Repository/SecondaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SecondaryRepository extends JpaRepository<SecondaryEntity, Long> {
+public interface SecondaryRepository extends JpaRepository<SecondaryEntity, Integer> {
 }
